Initial Mod Refactor
Refactor Mod for easier development / maintenance in future.
