+++ v1/core/StationManager.cs
@@ -1,0 +1,77 @@
+﻿using Model;
+using RollingStock;
+using System.Collections.Generic;
+using System.Linq;
+
+
+namespace RouteManager
+{
+    public static class StationManager
+    {
+        //private static Dictionary<string, bool> stationSelections = new Dictionary<string, bool>();
+
+        //Define 3d space coordinates of each viable station
+        public static Dictionary<string, StationData> Stations = new Dictionary<string, StationData>
+        {
+            { "sylva", new StationData(24634.5f, 620.57f, -941.23f, 24563.24f, 620.57f, -935.94f, 24598.87f, 620.57f, -938.585f, 71.45608232f) },
+            { "dillsboro", new StationData(22379.87f, 603.17f, -1410.88f, 22326.76f, 603.17f, -1434.12f, 22353.315f, 603.17f, -1422.5f, 57.9721459f) },
+            { "wilmot", new StationData(16511.31f, 569.97f, 2326.23f, 16493.52f, 569.97f, 2329.14f, 16502.415f, 569.97f, 2327.685f, 18.0264306f) },
+            { "whittier", new StationData(12267.1f, 561.45f, 5864.33f, 12279.19f, 561.45f, 5893.68f, 12273.145f, 561.45f, 5879.005f, 31.74256763f) },
+            { "ela", new StationData(9569.54f, 546.61f, 7404.1f, 9554.41f, 546.61f, 7409.92f, 9561.975f, 546.61f, 7407.01f, 16.21077728f) },
+            { "bryson", new StationData(4530.43f, 528.97f, 5428.56f, 4473.52f, 528.97f, 5407.87f, 4501.975f, 528.97f, 5418.215f, 60.55430786f) },
+            { "hemingway", new StationData(2820.64f, 578.52f, 3079.64f, 2815.72f, 578.54f, 3055.54f, 2818.18f, 578.53f, 3067.59f, 24.59708926f) },
+            { "alarkajct", new StationData(1745.6f, 590.23f, 1503.32f, 1737.93f, 589.78f, 1425.91f, 1741.765f, 590.005f, 1464.615f, 77.79035609f) },
+            { "cochran", new StationData(1996.88f, 591.62f, -205.13f, 2007.29f, 591.85f, -218.98f, 2002.085f, 591.735f, -212.055f, 17.32753589f) },
+            { "alarka", new StationData(4170.52f, 644.81f, -3113.05f, 4201.17f, 645.24f, -3140.48f, 4185.845f, 645.025f, -3126.765f, 41.13407711f) },
+            {"alarkajctn", new StationData(1738.56f, 590.22f, 1504.86f, 1713.16f, 589.78f, 1431.7f, 1725.86f, 590f, 1468.28f, 77.44507215f)  },
+            { "almond", new StationData(-6340.3f, 524.97f, -1291.01f, -6316.44f, 524.97f, -1347.1f, -6328.37f, 524.97f, -1319.055f, 60.95398018f) },
+            { "nantahala", new StationData(-15594.29f, 595.2f, -10588.8f, -15642.63f, 595.51f, -10646.29f, -15618.46f, 595.355f, -10617.545f, 75.11292698f) },
+            { "topton", new StationData(-18969.52f, 793.22f, -15217.75f, -18977.49f, 792.7f, -15231.27f, -18973.505f, 792.96f, -15224.51f, 15.70292011f) },
+            { "rhodo", new StationData(-22993.12f, 653.53f, -18005.08f, -23014.11f, 653.15f, -18030.5f, -23003.615f, 653.34f, -18017.79f, 32.96818011f) },
+            { "andrews", new StationData(-29923.78f, 538.97f, -20057.8f, -29990.74f, 538.97f, -20092.33f, -29957.26f, 538.97f, -20075.065f, 75.33898393f) }
+        };
+
+        //Determine if station is selected
+        public static bool IsStationSelected(PassengerStop stop, Car locomotive)
+        {
+            return LocoTelem.UIStationSelections[locomotive].TryGetValue(stop.identifier, out bool isSelected) && isSelected;
+        }
+
+        //Update station selection
+        public static void SetStationSelected(PassengerStop stop, Car locomotive, bool isSelected)
+        {
+            LocoTelem.UIStationSelections[locomotive][stop.identifier] = isSelected;
+        }
+
+        //Check if Consist has any stations enabled
+        public static bool IsAnyStationSelectedForLocomotive(Car locomotive)
+        {
+            // Check if the locomotive exists in the SelectedStations dictionary
+            if (LocoTelem.SelectedStations.TryGetValue(locomotive, out List<PassengerStop> selectedStations))
+            {
+                // Return true if there is at least one selected station
+                return selectedStations.Any();
+            }
+
+            // Return false if the locomotive is not found or no stations are selected
+            return false;
+        }
+
+        //Setup create station selection list of current consist.
+        public static void InitializeStationSelectionForLocomotive(Car locomotive)
+        {
+            if (!LocoTelem.UIStationSelections.ContainsKey(locomotive))
+            {
+                var stationSelectionsForLocomotive = new Dictionary<string, bool>();
+                var allStops = PassengerStop.FindAll();
+
+                foreach (var stop in allStops)
+                {
+                    stationSelectionsForLocomotive[stop.identifier] = false;
+                }
+
+                LocoTelem.UIStationSelections[locomotive] = stationSelectionsForLocomotive;
+            }
+        }
+    }
+}
